feat(outputs): implement outputs

+++ src/main/java/fr/ing/interview/bankaccountkata/service/AccountServiceInterface.java
@@ -2,8 +2,10 @@ package fr.ing.interview.bankaccountkata.service;
 
 import fr.ing.interview.bankaccountkata.entity.Account;
 
+import java.math.BigDecimal;
+
 public interface AccountServiceInterface {
-    Account updateBalance(Long accountId, int amount);
+    Account updateBalance(Long accountId, BigDecimal amount);
 
     Account findAccountById(Long accountId) throws Exception;
 }
